pyproject.toml
[project]
name = "lchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    # llama ccp <local ai model>
    "langchain-community>=0.4.1",
    "llama-cpp-python>=0.3.20",
    # llama ccp </local ai model>
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = ["pytest>=9.0.3", "ruff>=0.15.10", "ty>=0.0.29"]
